Python/requirements.txt
numpy
matplotlib
pyserial

